+++ persistent/test/main.hs
@@ -30,10 +30,9 @@ import qualified Database.Persist.THSpec as THSpec
 
 main :: IO ()
 main = hspec $ do
-    describe "Database.Persist" $ do
-        describe "THSpec" THSpec.spec
+    describe "Database" $ describe "Persist" $ do
+        THSpec.spec
 
-    THSpec.spec
     describe "splitExtras" $ do
         let helloWorldTokens = Token "hello" :| [Token "world"]
             foobarbazTokens = Token "foo" :| [Token "bar", Token "baz"]